# hermes_wisdom/contracts/gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gateway-gateway connector API",
    "version": "1.0.0",
    "description": "The Hermes Agent relay/passthrough connector's own HTTP API: health, relay bootstrap (enroll/provision/policy), the management plane (link/scope/principal/dm-default/deprovision), and connector-hosted OAuth onboarding. Does NOT cover inbound provider webhook ingress (Discord/Slack/Meta), which uses vendor-defined signed payloads."
  },
  "servers": [
    {
      "url": "https://gateway-gateway.staging-nousresearch.com",
      "description": "Staging"
    },
    {
      "url": "https://gateway-gateway.nousresearch.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A NAS-signed agent-aud token (managed) OR the instance's per-gateway secret (self-hosted)."
      },
      "tenantAdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "A server-verified tenant identity token carrying a canonical subject and an org role/group claim that the deployment's role authority maps to authoritative admin (NAS OWNER/ADMIN by default; enterprise deployments pin their own admin allowlist). Connector-admin, agent-aud, and gateway-secret credentials are not accepted."
      },
      "oauthStartToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Short-lived agent-aud token authenticating the instance."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable failure reason.",
            "example": "platform is required"
          },
          "detail": {
            "type": "string",
            "description": "Optional bounded diagnostic safe to display to the authenticated caller."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WisdomInstallationDeactivateResponse": {
        "type": "object",
        "properties": {
          "skill_id": {
            "type": "string"
          },
          "installation_id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "inactive"
            ]
          }
        },
        "required": [
          "skill_id",
          "installation_id",
          "state"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/sync/wisdom/skills": {
      "get": {
        "summary": "Discovery: list the org's active shared skills.",
        "description": "Attribution is by user ID only — display-name joins happen in the NAS BFF (#211 R9), keeping the plane free of NAS user-profile reads.",
        "tags": [
          "wisdom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived",
                "taken_down"
              ],
              "description": "Lifecycle state. Inactive listings require wisdom:admin and OWNER/ADMIN."
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Opaque continuation cursor returned by the previous page."
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "description": "Page size from 1 through 100 (default 50).",
              "example": "50"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Active skills with latest version, lifecycle fence, author copy, server facts, and install counts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WisdomSkillListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing wisdom:read scope, NAS-admin claim, or org_role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sync/wisdom/skills/{id}": {
      "get": {
        "summary": "Skill detail + the authoritative version index (R4).",
        "tags": [
          "wisdom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Draft or skill id (uuid).",
              "example": "6a1f6f3e-7f7d-4e57-9b6e-2f6f0d3c9a11"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Skill lifecycle fence + append-only version history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WisdomSkillDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing wisdom:read scope, NAS-admin claim, or org_role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Skill not found in the token's org.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sync/wisdom/installations/{installationId}/skills/{skillId}": {
      "delete": {
        "summary": "Idempotently deactivate one managed installation.",
        "description": "Preserves the installation and append-only audit history. Replaying an uninstall after a crash returns the same inactive outcome.",
        "tags": [
          "wisdom"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Random revocable installation id (client-minted).",
              "example": "inst-5f0c9c2a7b1e"
            },
            "required": true,
            "name": "installationId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Stable Wisdom skill id."
            },
            "required": true,
            "name": "skillId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Installation is inactive (including idempotent replay).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WisdomInstallationDeactivateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid bearer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Missing wisdom:install scope, NAS-admin claim, or owned installation identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}